refactor: group the implementations
