+++ lib/implementations/indexeddb.js
@@ -1,7 +1,7 @@
-import { executeCallback } from './utils';
+import { executeCallback } from './../utils';
 import IDBKeyRange from './idbKeyRange';
 
-export function startsWithIndexedDB(prefix, callback) {
+export function startsWith(prefix, callback) {
     var localforageInstance = this;
     var promise = new Promise(function(resolve, reject) {
         localforageInstance.ready().then(function() {
