+++ lib/implementations/websql.js
@@ -1,6 +1,6 @@
-import { getSerializerPromise, executeCallback } from './utils';
+import { getSerializerPromise, executeCallback } from './../utils';
 
-export function startsWithWebsql(prefix, callback) {
+export function startsWith(prefix, callback) {
     var localforageInstance = this;
     var promise = new Promise(function(resolve, reject) {
         localforageInstance.ready().then(function() {
