+++ lib/localforage-startswith.js
@@ -1,38 +1,32 @@
 import localforage from 'localforage';
-import { executeCallback, getItemKeyValue } from './utils';
-import { startsWithIndexedDB } from './startswith-indexeddb';
-import { startsWithWebsql } from './startswith-websql';
+import { startsWith as startsWithIndexedDB } from './implementations/indexeddb';
+import { startsWith as startsWithWebsql } from './implementations/websql';
+import { startsWith as startsWithGeneric } from './implementations/generic';
 
-export function startsWithGeneric(prefix, callback) {
-    var localforageInstance = this;
-    var promise = new Promise(function(resolve, reject) {
-        localforageInstance.keys().then(function(keys) {
+export { startsWith as startsWithGeneric } from './implementations/generic';
 
-            var itemPromises = [];
+// export function keysStartingWithGeneric(prefix, callback) {
+//     var localforageInstance = this;
+//     var promise = new Promise(function(resolve, reject) {
+//         localforageInstance.keys().then(function(keys) {
 
-            var prefixLength = prefix.length;
-            for (var i = 0, len = keys.length; i < len; i++) {
-                var key = keys[i];
+//             var resultKeys = [];
 
-                if (key.slice(0, prefixLength) === prefix) {
-                    itemPromises.push(getItemKeyValue.call(localforageInstance, key));
-                }
-            }
+//             var prefixLength = prefix.length;
+//             for (var i = 0, len = keys.length; i < len; i++) {
+//                 var key = keys[i];
 
-            Promise.all(itemPromises).then(function(keyValuePairs) {
-                var result = {};
-                for (var i = 0, len = keyValuePairs.length; i < len; i++) {
-                    var keyValuePair = keyValuePairs[i];
+//                 if (key.slice(0, prefixLength) === prefix) {
+//                     resultKeys.push(key);
+//                 }
+//             }
 
-                    result[keyValuePair.key] = keyValuePair.value;
-                }
-                resolve(result);
-            }).catch(reject);
-        }).catch(reject);
-    });
-    executeCallback(promise, callback);
-    return promise;
-}
+//             resolve(resultKeys);
+//         }).catch(reject);
+//     });
+//     executeCallback(promise, callback);
+//     return promise;
+// }
 
 export function localforageStartsWith(prefix, callback) {
     var localforageInstance = this;
@@ -47,10 +41,24 @@ export function localforageStartsWith(prefix, callback) {
     }
 }
 
+// export function localforageKeysStartingWith(prefix, callback) {
+//     var localforageInstance = this;
+//     var currentDriver = localforageInstance.driver();
+
+//     if (currentDriver === localforageInstance.INDEXEDDB) {
+//         return keysStartingWithIndexedDB.call(localforageInstance, prefix, callback);
+//     } else if (currentDriver === localforageInstance.WEBSQL) {
+//         return keysStartingWithWebsql.call(localforageInstance, prefix, callback);
+//     }else {
+//         return keysStartingWithGeneric.call(localforageInstance, prefix, callback);
+//     }
+// }
+
 export function extendPrototype(localforage) {
     var localforagePrototype = Object.getPrototypeOf(localforage);
     if (localforagePrototype) {
         localforagePrototype.startsWith = localforageStartsWith;
+        // localforagePrototype.keysStartingWith = localforageKeysStartingWith;
     }
 }
